Rename all h files to match c++ convention
Also update the cmakelist file to reflect this changes.
