+++ include/dog.hpp
@@ -1,5 +1,5 @@
-#ifndef DOG_H
-#define DOG_H
+#ifndef DOG_HPP
+#define DOG_HPP
 
 #include <opencv2/opencv.hpp>
