+++ include/kuwahara.hpp
@@ -1,0 +1,3 @@
+#ifndef KUWAHARA_HPP
+#define KUWAHARA_HPP
+#endif
